interpolation search

+++ 0x1E-search_algorithms/102-interpolation.c
@@ -12,8 +12,8 @@ int interpolation_search(int *array, size_t size, int value)
 {
 	size_t pos;
 	double c;
-	int low = 0;
-	int high = size - 1;
+	size_t low = 0;
+	size_t high = size - 1;
 
 	if (array == NULL)
 		return (-1);
